requirements.txt
python-telegram-bot
python-magic